+++ include/util/coordinate.hpp
@@ -33,6 +33,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <boost/numeric/conversion/cast.hpp>
 
 #include <cstddef>
+#include <cstdint>
 #include <iosfwd> //for std::ostream
 #include <sstream>
 #include <string>