+++ test/nodejs/index.js
@@ -118,10 +118,17 @@ test('constructor: takes a default_radius argument', function(assert) {
     assert.ok(osrm);
 });
 
+test('constructor: takes a default_radius unlimited argument', function(assert) {
+    assert.plan(1);
+    var osrm = new OSRM({algorithm: 'MLD', path: monaco_mld_path, default_radius: 'unlimited'});
+    assert.ok(osrm);
+});
+
 test('constructor: throws if default_radius is not a number', function(assert) {
-    assert.plan(2);
-    assert.throws(function() { new OSRM({algorithm: 'MLD', path: monaco_mld_path, default_radius: 'abc'}); }, /default_radius must be an integral number/, 'Does not accept string');
+    assert.plan(3);
+    assert.throws(function() { new OSRM({algorithm: 'MLD', path: monaco_mld_path, default_radius: 'abc'}); }, /default_radius must be unlimited or an integral number/, 'Does not accept invalid string');
     assert.ok(new OSRM({algorithm: 'MLD', path: monaco_mld_path, default_radius: 1}), 'Does accept number');
+    assert.ok(new OSRM({algorithm: 'MLD', path: monaco_mld_path, default_radius: 'unlimited'}), 'Does accept unlimited');
 });
 
 test('constructor: parses custom limits', function(assert) {
@@ -135,6 +142,7 @@ test('constructor: parses custom limits', function(assert) {
         max_locations_map_matching: 1,
         max_results_nearest: 1,
         max_alternatives: 1,
+        default_radius: 1
     });
     assert.ok(osrm);
 });
@@ -150,9 +158,47 @@ test('constructor: throws on invalid custom limits', function(assert) {
             max_locations_distance_table: false,
             max_locations_map_matching: 'a lot',
             max_results_nearest: null,
-            max_alternatives: '10'
+            max_alternatives: '10',
+            default_radius: '10'
         })
     });
+});
+test('constructor: throws on invalid disable_feature_dataset option', function(assert) {
+    assert.plan(1);
+    assert.throws(function() {
+        var osrm = new OSRM({
+            path: monaco_path,
+            disable_feature_dataset: ['NOT_EXIST'],
+        })
+    });
+});
+
+test('constructor: throws on non-array disable_feature_dataset', function(assert) {
+    assert.plan(1);
+    assert.throws(function() {
+        var osrm = new OSRM({
+            path: monaco_path,
+            disable_feature_dataset: 'ROUTE_GEOMETRY',
+        })
+    });
+});
+
+test('constructor: ok on valid disable_feature_dataset option', function(assert) {
+    assert.plan(1);
+    var osrm = new OSRM({
+        path: monaco_path,
+        disable_feature_dataset: ['ROUTE_GEOMETRY'],
+    });
+    assert.ok(osrm);
+});
+
+test('constructor: ok on multiple overlapping disable_feature_dataset options', function(assert) {
+    assert.plan(1);
+    var osrm = new OSRM({
+        path: monaco_path,
+        disable_feature_dataset: ['ROUTE_GEOMETRY', 'ROUTE_STEPS'],
+    });
+    assert.ok(osrm);
 });
 
 require('./route.js');